% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winner.R
\name{bowling_leaderboard}
\alias{bowling_leaderboard}
\title{Determine the winner of a bowling game between two players.}
\usage{
bowling_leaderboard(player1_rolls, player2_rolls)
}
\arguments{
\item{player1_rolls}{A numeric vector representing the scores for every throw of player 1.}

\item{player2_rolls}{A numeric vector representing the scores for every throw of player 2.}
}
\value{
A character string indicating the winner of the game ("Player 1", "Player 2", or "Tie").
}
\description{
This function calculates the scores for two players in a bowling game and determines the winner. It also displays the final scores of both players.
}
\examples{
player1_rolls <- c(2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2)
player2_rolls <- c(2,2,2,2,8,2,6,2,2,2,2,2,5,2,2,2,4,2,2,3)
bowling_leaderboard(player1_rolls, player2_rolls)

}
