% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winner.R
\name{bowling_winner}
\alias{bowling_winner}
\title{Determine the winner of a bowling game between two players.}
\usage{
bowling_winner(player_rolls)
}
\arguments{
\item{player_rolls}{A numeric vector representing the scores for every throw of every player}
}
\value{
A character string indicating the winner of the game ("Player 1", "Player 2", or "Tie").
}
\description{
This function calculates the scores for two players in a bowling game and determines the winner. It also displays the final scores of both players.
}
\examples{

player_rolls <- player_throws(player_no = c(1, 2, 3),
throws = list(
c(2, 3, 4, 5, 6, 4, 8, 1, 1, 2, 3, 4, 5, 5, 7, 2, 9, 1, 4, 6, 10),
c(2, 3, 4, 5, 6, 4, 8, 1, 1, 2, 3, 4, 5, 5, 7, 2, 9, 1, 4, 3),
c(1, 2, 3, 4, 5, 3, 1, 8, 9, 1, 2, 3, 4, 5, 2, 7, 8, 1, 10, 0, 10)
))
bowling_winner(player_rolls)

}
