% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlingpin-package.R
\docType{package}
\name{bowlingpin-package}
\alias{bowlingpin-package}
\title{bowlingpin: Counting the score of a ten-pin bowling game}
\description{
It counts the score of a ten-pin bowling game for two players and returns the final score after the players add their score. For now, it just adds up the score and might give the exact score-sum of strikes and spares.
}
\author{
\strong{Maintainer}: Disha Rathod \email{drat0009@student.monash.edu}

}
\keyword{internal}
