% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlingpin.R
\name{bowlingpin}
\alias{bowlingpin}
\title{Calculate Bowling Pin Scores}
\usage{
bowlingpin(player_rolls)
}
\arguments{
\item{player_rolls}{A record containing player numbers and their throws.}
}
\value{
A record with players' scores calculated.
}
\description{
This function calculates the total scores for each player's throws in a bowling game and prints it in custom vector format.
}
\examples{
player_rolls <- player_throws(player_no = c(1, 2, 3),
  throws = list(
  c(2, 3, 4, 5, 6, 4, 8, 1, 1, 2, 3, 4, 5, 5, 7, 2, 9, 1, 4, 6, 10),
  c(2, 3, 4, 5, 6, 4, 8, 1, 1, 2, 3, 4, 5, 5, 7, 2, 9, 1, 4, 3),
  c(1, 2, 3, 4, 5, 3, 1, 8, 9, 1, 2, 3, 4, 5, 2, 7, 8, 1, 10, 0, 10)
   ))

bowlingpin(player_rolls)

}
