% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlingpin.R
\name{calculate_bowling_score}
\alias{calculate_bowling_score}
\title{Calculate the sum of throws}
\usage{
calculate_bowling_score(throws)
}
\arguments{
\item{player_rolls}{A record containing player numbers and their throws.}
}
\value{
A record with players' scores calculated.
}
\description{
This function calculates the total scores for each player's throws in a bowling game and returns a data frame.
}
\keyword{internal}
