% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxgutters.R
\name{calculate_gutters}
\alias{calculate_gutters}
\title{Calculate the number of gutters in a player's rolls.}
\usage{
calculate_gutters(player_rolls)
}
\arguments{
\item{player_rolls}{A numeric vector representing the scores for every throw of a player.}
}
\value{
An integer representing the number of gutters in the player's rolls.
}
\description{
Calculate the number of gutters in a player's rolls.
}
\examples{
player_rolls <- c(4, 5, 0, 6, 0, 0, 0, 0, 0, 0, 7, 3, 0, 0, 5, 0, 5, 0, 10, 0, 5)
calculate_gutters(player_rolls)
}
