% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxstrikes.R
\name{calculate_strikes}
\alias{calculate_strikes}
\title{Calculate the number of strikes in a player's rolls.}
\usage{
calculate_strikes(player_rolls)
}
\arguments{
\item{player_rolls}{A numeric vector representing the scores for every throw of a player.}
}
\value{
An integer indicating the number of strikes in the player's rolls.
}
\description{
This function counts the number of strikes (rolls with a score of 10) in a player's rolls.
}
\keyword{internal}
