% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_throws.R
\name{format.throws}
\alias{format.throws}
\title{Custom Formatting Function for Bowling Scorecards}
\usage{
\method{format}{throws}(x, ...)
}
\arguments{
\item{x}{A record containing player numbers, their throws, and scores.}

\item{...}{Additional parameters (not used).}
}
\value{
A formatted representation of the bowling scorecards.
}
\description{
This function formats the bowling scorecards for each player in a custom vector format.
}
