% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxgutters.R
\name{max_gutters}
\alias{max_gutters}
\title{Determine the player with the most gutters in a game.}
\usage{
max_gutters(player1_rolls, player2_rolls)
}
\arguments{
\item{player1_rolls}{A numeric vector representing the scores for every throw of player 1.}

\item{player2_rolls}{A numeric vector representing the scores for every throw of player 2.}
}
\value{
A message indicating which player had the most gutters and the number of gutters they had.
}
\description{
This function compares the number of gutters for two players and determines which player had the most gutters in the game.
}
\examples{
player1_rolls <- c(0, 0, 3, 0, 0, 0, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
player2_rolls <- c(0, 0, 0, 0, 0, 0, 0, 0, 6, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0)
max_gutters(player1_rolls, player2_rolls)

}
