% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxstrikes.R
\name{max_strikes}
\alias{max_strikes}
\title{Determine the player with the most strikes in the game.}
\usage{
max_strikes(player1_rolls, player2_rolls)
}
\arguments{
\item{player1_rolls}{A numeric vector representing the scores for every throw of player 1.}

\item{player2_rolls}{A numeric vector representing the scores for every throw of player 2.}
}
\value{
NULL (prints the player with the most strikes and their corresponding number of strikes).
}
\description{
This function compares the number of strikes between two players and identifies the player with the highest number of strikes.
}
\examples{
player1_rolls <- c(10, 3, 4, 10, 5, 10, 7, 2, 3, 4, 10, 10, 10, 6, 7, 8, 9, 1, 2, 3)
player2_rolls <- c(10, 10, 4, 3, 5, 10, 10, 2, 3, 4, 10, 10, 10, 6, 7, 8, 9, 1, 2, 3)
max_strikes(player1_rolls, player2_rolls)

}
