% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_throws.R
\name{player_throws}
\alias{player_throws}
\title{Player Throws and Scores}
\usage{
player_throws(player_no, throws, score = numeric())
}
\arguments{
\item{player_no}{Player number}

\item{throws}{A list of throws of each player}

\item{score}{Total score}
}
\value{
Throws a record
}
\description{
Functions to handle player throws in a ten pin bowling game and calculate their scores.
}
\examples{

player_rolls <- player_throws(player_no = c(1, 2, 3),
throws = list(
c(2, 3, 4, 5, 6, 4, 8, 1, 1, 2, 3, 4, 5, 5, 7, 2, 9, 1, 4, 6, 10),
c(2, 3, 4, 5, 6, 4, 8, 1, 1, 2, 3, 4, 5, 5, 7, 2, 9, 1, 4, 3),
c(1, 2, 3, 4, 5, 3, 1, 8, 9, 1, 2, 3, 4, 5, 2, 7, 8, 1, 10, 0, 10)
))
}
