% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_throws.R
\name{vec_cast.data.frame.throws}
\alias{vec_cast.data.frame.throws}
\title{Vectorized Casting Function for Data Frame to Throws}
\usage{
\method{vec_cast}{data.frame.throws}(x, to, ...)
}
\arguments{
\item{x}{A data frame to be casted.}

\item{to}{The desired format for the casting.}

\item{...}{Additional arguments to be passed.}
}
\value{
A casted version of the data frame in throws format.
}
\description{
This function takes a data frame and casts it into a throws format.
}
