% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_throws.R
\name{vec_cast.throws.data.frame}
\alias{vec_cast.throws.data.frame}
\title{Vectorized Casting Function for Throws Data Frame}
\usage{
\method{vec_cast}{throws.data.frame}(x, to, ...)
}
\arguments{
\item{x}{A data frame containing player number, throws, and score.}

\item{to}{The desired format for the casting.}

\item{...}{Additional arguments to be passed.}
}
\value{
A casted version of the throws data frame.
}
\description{
This function takes a throws data frame and casts it into another form.
}
