% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_throws.R
\name{vec_ptype2.data.frame.throws}
\alias{vec_ptype2.data.frame.throws}
\title{Prototyping vector as data frame}
\usage{
vec_ptype2.data.frame.throws(x, y, ...)
}
\arguments{
\item{x}{First parameter of data frame class}

\item{y}{Second parameter of throws class}

\item{...}{Additional parameters}
}
\value{
A data frame type
}
\description{
Prototyping vector as data frame
}
