% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_throws.R
\name{vec_ptype2.throws.data.frame}
\alias{vec_ptype2.throws.data.frame}
\title{Prototyping as data frame in reverse order}
\usage{
vec_ptype2.throws.data.frame(x, y, ...)
}
\arguments{
\item{x}{First parameter of data frame class}

\item{y}{Second parameter of throws class}

\item{...}{Additional parameters}
}
\value{
A data frame type
}
\description{
Prototyping as data frame in reverse order
}
