% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlingpin.R
\name{bowlingpin}
\alias{bowlingpin}
\title{Main function to run the ten-pin bowling game for two players}
\usage{
bowlingpin(player_rolls)
}
\arguments{
\item{player1_rolls}{A numeric vector having scores for every throw of player 1}

\item{player2_rolls}{A numeric vector having scores for every throw of player 2}
}
\value{
A numeric vector containing the total score for each user.
}
\description{
Enter your individual game scores and the function will give you the total score at the end
}
\examples{
player_rolls<- player_throws(player_no = c(1, 2, 3),
throws = list(
c(2, 3, 4, 5, 6, 4, 8, 1, 1, 2, 3, 4, 5, 5, 7, 2, 9, 1, 4, 6, 10),
c(2, 3, 4, 5, 6, 4, 8, 1, 1, 2, 3, 4, 5, 5, 7, 2, 9, 1, 4, 3),
c(1, 2, 3, 4, 5, 3, 1, 8, 9, 1, 2, 3, 4, 5, 2, 7, 8, 1, 10, 0, 10)
))
bowlingpin(player_rolls)
}
